requirements.txt
geopandas
matplotlib
scikit-learn
pyyaml
moviepy
bitstring
pygame
mapclassify
folium

